Remove unused code from previous implementation
